Refactor tests to match refactored code structure.
